+++ spotbugs-tests/src/test/java/edu/umd/cs/findbugs/detect/SharedVariableAtomicityDetectorTest.java
@@ -279,6 +279,14 @@ class SharedVariableAtomicityDetectorTest extends AbstractIntegrationTest {
     }
 
     @Test
+    void noBugNoCompoundOp() {
+        performAnalysis("multithreaded/compoundoperation/NoCompoundOp.class");
+        assertBugTypeCount(PRIMITIVE_BUG, 0);
+        assertBugTypeCount(WRITE_64BIT_BUG, 0);
+        assertBugTypeCount(OPS_BUG, 0);
+    }
+
+    @Test
     void noBugInnerLocalVarFields() {
         // Inner local vars
         performAnalysis("multithreaded/compoundoperation/InnerLocalVarFields.class");